% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFileInput.R
\name{selectFileInput}
\alias{selectFileInput}
\title{selectFileInput}
\usage{
selectFileInput(inputId, choices = NULL)
}
\arguments{
\item{inputId}{inputId used by Shiny.}

\item{choices}{Choices.}
}
\description{
creates a selectFileInput component
}
\details{
initialize a selectFileInput
}
