% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTableInput.R
\name{selectTableInput}
\alias{selectTableInput}
\title{selectTableInput}
\usage{
selectTableInput(
  inputId,
  headers = NULL,
  content = NULL,
  multiple = FALSE,
  selected = NULL,
  maxRows = 5,
  actionLink = NULL
)
}
\arguments{
\item{inputId}{inputId used by Shiny}

\item{headers}{The table headers (if NULL, content's column names will be used).}

\item{content}{The table content (data.frame like object).}

\item{multiple}{Is selection of multiple rows allowed?}

\item{selected}{Index of initial row(s) selected.}

\item{maxRows}{Max. number of rows to be displayed.}

\item{actionLink}{Label for actionLink. If not NULL, an extra column will be added with an actionLink.}
}
\description{
creates a selectTableInput component
}
\details{
initialize a selectTableInput
}
